requirements.txt
flask==3.0.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
python-docx==1.1.0
fpdf2==2.7.6
openai>=2.6.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
werkzeug==3.0.1
openpyxl==3.1.2
pyyaml==6.0.1
pillow==10.1.0
httpx>=0.28.1
scipy==1.16.2
